Input: overhaul touch controls with 1-finger pan, 2-finger box select, and 4-finger zoom
- Implement 1-finger smooth map scrolling with zero jump and 25px touch slop
- Add dedicated 2-finger box selection gesture with trailing finger shield
- Move map zoom to 4-finger pinch-to-zoom to avoid gesture conflicts
- Restore bottom-left command bar coordinates and mouse hover tooltips

+++ OpenRA.Mods.Common/Widgets/ProductionPaletteWidget.cs
@@ -556,7 +556,7 @@ namespace OpenRA.Mods.Common.Widgets
 			Game.Renderer.EnableAntialiasingFilter();
 			foreach (var icon in icons.Values)
 			{
-				var iconScale = icon.Sprite != null && icon.Sprite.Size.X > 0 ? (float)IconSize.X / icon.Sprite.Size.X : 1f;
+				var iconScale = icon.Sprite != null && icon.Sprite.Size.X > 0 ? IconSize.X / icon.Sprite.Size.X : 1f;
 				WidgetUtils.DrawSpriteCentered(icon.Sprite, icon.Palette, icon.Pos + iconOffset, iconScale);
 
 				// Draw the ProductionIconOverlay's sprites
@@ -572,12 +572,12 @@ namespace OpenRA.Mods.Common.Widgets
 							* (clock.CurrentSequence.Length - 1) / first.TotalTime);
 					clock.Tick();
 
-					var clockScale = clock.Image != null && clock.Image.Size.X > 0 ? (float)IconSize.X / clock.Image.Size.X : iconScale;
+					var clockScale = clock.Image != null && clock.Image.Size.X > 0 ? IconSize.X / clock.Image.Size.X : iconScale;
 					WidgetUtils.DrawSpriteCentered(clock.Image, icon.IconClockPalette, icon.Pos + iconOffset, clockScale);
 				}
 				else if (!buildableItems.Any(a => a.Name == icon.Name))
 				{
-					var cantBuildScale = cantBuild.Image != null && cantBuild.Image.Size.X > 0 ? (float)IconSize.X / cantBuild.Image.Size.X : iconScale;
+					var cantBuildScale = cantBuild.Image != null && cantBuild.Image.Size.X > 0 ? IconSize.X / cantBuild.Image.Size.X : iconScale;
 					WidgetUtils.DrawSpriteCentered(cantBuild.Image, icon.IconDarkenPalette, icon.Pos + iconOffset, cantBuildScale);
 				}
 			}
